Consolidate nav and auth modules
