+++ pages/index.tsx
@@ -1,15 +1,7 @@
-import { Container, Typography } from "@mui/material";
 import type { NextPage } from "next";
 
-const Home: NextPage = () => {
-  return (
-    <Container sx={{ flex: 1 }}>
-      <Typography variant="h4">Welcome to the new OBGS webapp</Typography>
-      <Typography variant="body1">
-        This app is a work in progress. Please check back later.
-      </Typography>
-    </Container>
-  );
-};
+import Home from "modules/home";
 
-export default Home;
+const HomePage: NextPage = () => <Home />;
+
+export default HomePage;
